Add silencing of missing far alerts
