+++ datascience/tests/test_data/remote_database/V666.18__Reset_test_silenced_alerts.sql
@@ -8,4 +8,9 @@ INSERT INTO silenced_alerts (
     'MYNAMEIS', 'ABC000658985', 'OHMYGOSH', 'OGMJ', 'INTERNAL_REFERENCE_NUMBER',
     NOW() + ('15 DAYS')::interval, NULL,
     '{"type": "THREE_MILES_TRAWLING_ALERT", "seaFront": "Facade B", "flagState": "FR"}'
+),
+(
+    'DEVINER FIGURE CONSCIENCE', 'ABC000542519', 'RO237719', 'FQ7058', 'INTERNAL_REFERENCE_NUMBER',
+    NOW() + ('15 DAYS')::interval, NULL,
+    '{"type": "MISSING_FAR_ALERT", "seaFront": "Facade A", "flagState": "FR"}'
 );
