+++ datascience/tests/test_data/remote_database/V666.1__Reset_test_positions.sql
@@ -21,4 +21,13 @@ INSERT INTO positions (
 (   13735518, NULL,                      'ZZTOPACDC',               NULL, 'ZZ000000', 'I DO 4H REPORT',             'FR',       'FR',         NULL,                NULL,        -53.4320000000000022, -5.55100000000000016,  1,     203,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '8 hours 10 minutes',       'VMS',         true,      false,      0.0,                           INTERVAL '4 hour',            0.4,           false,      INTERVAL '3 days 4 hours'),
 (   13738407, NULL,                      'ZZTOPACDC',               NULL, 'ZZ000000', 'I DO 4H REPORT',             'FR',       'FR',         NULL,                NULL,        -53.429000000000002,  -5.54499999999999993,  2,     225,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '4 hour 10 minutes',        'VMS',         true,      false,      0.0,                           INTERVAL '4 hour',            0.4,           false,      INTERVAL '3 days 8 hours'),
 (   13740935, NULL,                      'ZZTOPACDC',               NULL, 'ZZ000000', 'I DO 4H REPORT',             'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   1.5,   195,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '10 minutes',               'VMS',         true,      NULL,       0.0,                           INTERVAL '4 hour',            0.4,           false,      INTERVAL '3 days 12 hours'),
-(   13786523, 'ABC000658985',            'OHMYGOSH',                NULL, 'OGMJ',     'MYNAMEIS',                   'FR',       'FR',         NULL,                NULL,        49.6060000000000016,  -0.735999999999999988, 1.3,   652,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '10 hours',                 'VMS',         true,      NULL,       0.0,                           INTERVAL '4 hour',            0.7,           true,       INTERVAL '3 days 12 hours');
+(   13786523, 'ABC000658985',            'OHMYGOSH',                NULL, 'OGMJ',     'MYNAMEIS',                   'FR',       'FR',         NULL,                NULL,        49.6060000000000016,  -0.735999999999999988, 1.3,   652,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '10 hours',                 'VMS',         true,      NULL,       0.0,                           INTERVAL '4 hour',            0.7,           true,       INTERVAL '3 days 12 hours'),
+(   13786524, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     241,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '1 day 3 hours 30 minutes', 'VMS',         false,     false,      3500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day'),
+(   13786525, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   4,     225,    (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '1 day 3 hours',            'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 30 minutes'),
+(   13786526, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '1 day 2 hours 30 minutes', 'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour'),
+(   13786527, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '1 day',                       'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour'),
+(   13786528, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '20 hours',                    'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour'),
+(   13786529, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '16 hours',                    'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour'),
+(   13786530, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '12 hours',                    'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour'),
+(   13786531, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '4 hours',                     'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour'),
+(   13786532, 'ABC000542519',            'RO237719',                NULL, 'FQ7058',   'DEVINER FIGURE CONSCIENCE',  'FR',       'FR',         NULL,                NULL,        43.3239999999999981,  5.35899999999999999,   3,     56,     (NOW() AT TIME ZONE 'UTC')::TIMESTAMP - INTERVAL '8 hours',                     'VMS',         false,     false,      2500,                          INTERVAL '30 minutes',        2.69,          true,       INTERVAL '1 day 1 hour');
